scan-back/ml/requirements.txt
# Deep Learning Training Requirements for Chess Piece Detection
# Install with: pip install -r requirements.txt

# Core deep learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data augmentation
albumentations>=1.3.0
opencv-python>=4.8.0

# ONNX export and inference
onnx>=1.14.0
onnxruntime>=1.15.0
onnx-simplifier>=0.4.0

# Utilities
numpy>=1.24.0
Pillow>=10.0.0
pyyaml>=6.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress and console output
rich>=13.0.0
tqdm>=4.65.0

# Optional: GPU acceleration
# nvidia-cuda-runtime-cu11>=11.8  # For CUDA 11.8
# nvidia-cudnn-cu11>=8.9.0        # For cuDNN

# Optional: Better performance
# ninja  # For faster compilation
# tensorboard  # For training visualization
